TP5

+++ src/main/java/corp/sky/nano/ProductRepository.java
@@ -2,6 +2,6 @@ package corp.sky.nano;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<ProductEntity, Integer> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
 
 }
